Revert lines' vetices (float64 -> int)

+++ image.go
@@ -96,7 +96,7 @@ func (i *Image) DrawImage(image *Image, options *DrawImageOptions) (err error) {
 }
 
 // DrawLine draws a line.
-func (i *Image) DrawLine(x0, y0, x1, y1 float64, clr color.Color) error {
+func (i *Image) DrawLine(x0, y0, x1, y1 int, clr color.Color) error {
 	return i.DrawLines(&line{x0, y0, x1, y1, clr})
 }
